Resolved all merge conflicts between the cost aggregation API PR and develop:
- Renumbered cost_records migration from 000013 to 000016 (already existed in develop)
- Merged Project struct with both MaxBudget (PR) and CircuitBreaker fields (develop)
- Merged server.go/main.go to include both cost handlers and notification/epicRun handlers
- Merged gen_server.go to include cost API endpoints alongside develop additions
- Merged openapi.yaml to include both cost aggregation schema and develop additions
- Fixed mock implementations to satisfy updated port.CostRepository and port.RunRepository interfaces
- Removed duplicate cost_repository.go (develop version) in favor of complete cost_repo.go (PR)

+++ backend/queries/cost_records.sql
@@ -8,8 +8,8 @@ SELECT * FROM cost_records WHERE run_step_id = $1 LIMIT 1;
 
 -- name: SumCostByProject :one
 SELECT COALESCE(SUM(cost_usd), 0)::DECIMAL(10,6) AS total_cost,
-       COALESCE(SUM(tokens_input), 0)             AS total_input,
-       COALESCE(SUM(tokens_output), 0)            AS total_output
+       COALESCE(SUM(tokens_input), 0)::BIGINT     AS total_input,
+       COALESCE(SUM(tokens_output), 0)::BIGINT    AS total_output
 FROM cost_records
 WHERE project_id = $1 AND created_at >= $2;
